Day 5 Quiz Submitted.
Word Guessing Game
Write a program that takes two input items from the user.  The first item is the user's name.  The second item is the user's word guess.  The program takes a list of 10 numbers ad randomly chooses one word from the list for the user to guess.  If the user makes a correct guess the program acknowledges the correct guess to the user and exits the program.  Each time the user makes an incorrect guess, the program informs the user that the guess was incorrect and to guess again.